// src/schema/studentConductCharacterAssessments.schema.ts
import { z } from "zod";

export const studentConductCharacterAssessmentsFormSchema = z.object({
  typeofAssessments: z.string().nonempty("Type of assesments is required"),
  continuousAssessmentWeek: z.array(z.string()),
  location: z.string().nonempty("Location is required"),
  state: z.string().optional(),
  cities: z.string().optional(),
  mission: z.string().optional(),
  vision: z.string().optional(),
  coreValues: z.string().optional(),
  yearClass: z.string().nonempty("Year/Class is required"),
  schoolLevel: z.string().nonempty("School level is required"),
  subSchoolLevel: z.string().optional(),
  studentAge: z.string().optional(),
  classesSocioEconomic: z.string().optional(),
  aim: z.string().optional(),
  term: z.string().nonempty("Term is required"),
  termTheme: z.string().optional(),
  week: z.string().nonempty("Week is required"),
  KPI: z.string().nonempty("KPI is required"),
  bloomLevel: z.string().nonempty("Bloom’s level is required"),
  technologyAccess: z.string().nonempty("Technology access is required"),
  studentConductLessonObjectives: z.string().nonempty("Lesson objectives is required"),
  subject: z.string().nonempty("Subject is required"),
  subjectLearning: z.string().optional(),
  teachingAids: z
    .array(z.string())
    .min(1, "Select at least one teaching aid"),
  classSize: z.string().nonempty("Class size is required"),
  timeAvailable: z.string().nonempty("Time available is required"),
  // sen
  sen: z.array(z.string()).optional(),
  noStudents: z.array(z.string()).optional(),
  security: z.array(z.string()).optional(),
  support: z.string().optional(),
  communicationMethod: z.string().optional(),
  mobility: z.string().optional(),
  sensoryConsideration: z.string().optional(),
  socialInteraction: z.string().optional(),
  cognitiveProcessingTime: z.string().optional(),
  medicalEmergencyProtocol: z.string().optional(),
  iepPlan: z.string().optional(),
  senOptions: z.array(z.string()).optional(),
  //assessment
  cbtTest: z.string().optional(),
  nationalTest: z.string().optional(),
  noQuestions: z.string().nonempty('Number of questions is required'),
  questionTypes: z.array(z.string()).nonempty("Select at least one question type").min(1, "Select at least one question type"),
  maxOptions: z.string().optional(),
  maxAnswers: z.number().min(1, "Select at least one question type").optional(),
  correctModel: z.string().optional(),
  explanationCorrectModel: z.string().optional(),
  assessmentWeight: z.string().optional(),
  assessmentLearning: z.string().optional(),
  submissionFormat: z.array(z.string()).optional(),
})

export type IStudentConductCharacterAssessmentsForm = z.infer<typeof studentConductCharacterAssessmentsFormSchema>